+++ tests/unit/test_cli.py
@@ -10,7 +10,7 @@ from libera_utils.io import packet_ingest
 @pytest.mark.parametrize(
     ("cli_args", "parsed"),
     [
-        (['packet-ingest', 'fakedir.json'],
+        (['packet-ingest', '--delete', 'fakedir.json'],
          argparse.Namespace(
              func=packet_ingest.ingest,
              manifest_filepath='fakedir.json',